refactor: move remove_final_predicates to working nogood
